+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Changed
+
+- update `stac-fastapi-*` version requirements to `>=5.2,<6.0`
+- add pgstac health-check in `/_mgmt/health`
+
 ## [5.0.2] - 2025-04-07
 
 ### Fixed